Auditing funktioniert fürs erste wieder wie vorher. leistungsoptimierung durch producer-consumer-pattern folgt.
